Changed json @class property to @type property to handle polymorphic types using class type. Catching the exceptions using multi-catch statement in main. Created a resources folder in main directory and moved jsonData there. Changed the name of json file and added ".json" extension. The name is now "ShapesData.json"
