+++ resources/ShapesData.json
@@ -1,0 +1,18 @@
+{
+  "shapes": [
+    {
+      "@type": "circle",
+      "radius": 5
+    },
+    {
+      "@type": "Pentagon",
+      "apothem": 3,
+      "side": 2
+    },
+    {
+      "@type": "rectangle",
+      "length": 20.0,
+      "width": 25.0
+    }
+  ]
+}
