--- src/main/java/com/ebricks/shape/data/JsonData
@@ -1,18 +1,0 @@
-{
-  "shapes": [
-    {
-      "@class": "com.ebricks.shape.shapes.Circle",
-      "radius": 5
-    },
-    {
-      "@class": "com.ebricks.shape.shapes.Pentagon",
-      "apothem": 3,
-      "side": 2
-    },
-    {
-      "@class": "com.ebricks.shape.shapes.Rectangle",
-      "length": 20.0,
-      "width": 25.0
-    }
-  ]
-}
