+++ src/main/java/com/ebricks/shape/main/Driver.java
@@ -1,30 +1,31 @@
 package com.ebricks.shape.main;
 
 import com.ebricks.shape.processor.ShapeProcessor;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
 import java.util.concurrent.ExecutionException;
 
+
 public class Driver {
 
-    public static void main(String[] args){
+    public static Logger logger = LogManager.getRootLogger();
+
+    public static void main(String[] args) {
 
         ShapeProcessor shapeProcessor = new ShapeProcessor();
 
         try {
             shapeProcessor.init();
             shapeProcessor.process();
-        } catch (IOException e) {
-            ShapeProcessor.logger.error(e);
-        } catch (InterruptedException e) {
-            ShapeProcessor.logger.error(e);
-        } catch (ExecutionException e) {
-            ShapeProcessor.logger.error(e);
-        }finally {
+        } catch (IOException | InterruptedException | ExecutionException e) {
+            logger.error(e);
+        } finally {
             try {
                 shapeProcessor.end();
             } catch (InterruptedException e) {
-                ShapeProcessor.logger.error(e);
+                logger.error(e);
             }
         }
