+++ src/main/java/com/ebricks/shape/model/Rectangle.java
@@ -1,8 +1,6 @@
-package com.ebricks.shape.shapes;
+package com.ebricks.shape.model;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonGetter;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.*;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
