+++ src/main/java/com/ebricks/shape/model/Shape.java
@@ -1,0 +1,19 @@
+package com.ebricks.shape.model;
+
+import com.fasterxml.jackson.annotation.JsonSubTypes;
+import com.fasterxml.jackson.annotation.JsonTypeInfo;
+
+@JsonTypeInfo(use = JsonTypeInfo.Id.NAME)
+@JsonSubTypes({
+        @JsonSubTypes.Type(value = Circle.class),
+        @JsonSubTypes.Type(value = Rectangle.class, name = "rectangle"),
+        @JsonSubTypes.Type(value = Pentagon.class)
+})
+public abstract class Shape {
+
+    public abstract double area();
+
+    public abstract double perimeter();
+
+    public abstract Shape display();
+}
