+++ src/main/java/com/ebricks/shape/processor/ShapeProcessor.java
@@ -1,8 +1,8 @@
 package com.ebricks.shape.processor;
 
-import com.ebricks.shape.shapes.Canvas;
+import com.ebricks.shape.model.Canvas;
 import com.ebricks.shape.executor.ShapeExecutor;
-import com.ebricks.shape.shapes.Shape;
+import com.ebricks.shape.model.Shape;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -24,18 +24,19 @@ public class ShapeProcessor {
     public void init() throws IOException {
 
         objectMapper = new ObjectMapper();
-        canvasReader = (Canvas) objectMapper.readValue(new File("src\\main\\java\\com\\ebricks\\shape\\data\\JsonData"), Canvas.class);
+        canvasReader = (Canvas) objectMapper.readValue(new File("resources/ShapesData.json"), Canvas.class);
 
         service = Executors.newFixedThreadPool(2);
 
         futures = new ArrayList<Future>();
+    }
+
+    public void process() throws ExecutionException, InterruptedException {
+
         for (Shape shape : canvasReader.getShapes()) {
             ShapeExecutor shapeExecutor = new ShapeExecutor(shape);
             futures.add(service.submit(shapeExecutor));
         }
-    }
-
-    public void process() throws ExecutionException, InterruptedException {
 
         for (Future<Shape> future : futures) {
             Shape shape = future.get();
@@ -45,15 +46,10 @@ public class ShapeProcessor {
 
     public void end() throws InterruptedException {
 
-        if(canvasReader != null){
-            canvasReader.getShapes().clear();
-        }
         if (service != null) {
             service.shutdown();
             service.awaitTermination(10, TimeUnit.SECONDS);
         }
-        if (futures != null) {
-            futures.clear();
-        }
+
     }
 }
